Rename Pregel/Graph/ to Pregel/GraphFormat to avoid ambiguities
