+++ arangod/Pregel/Algos/DMID/DMIDMessageFormat.h
@@ -28,7 +28,7 @@
 #pragma once
 
 #include "Pregel/CommonFormats.h"
-#include "Pregel/Graph/Graph.h"
+#include "Pregel/GraphStore/Graph.h"
 #include "Pregel/GraphFormat.h"
 #include "Pregel/MessageFormat.h"
