+++ arangod/Pregel/CommonFormats.h
@@ -29,7 +29,7 @@
 
 #include <map>
 
-#include "Pregel/Graph/Graph.h"
+#include "Pregel/GraphStore/Graph.h"
 #include "Pregel/GraphFormat.h"
 #include "Pregel/MessageFormat.h"
 #include "Pregel/VertexComputation.h"
