+++ arangod/Pregel/VertexComputation.h
@@ -26,11 +26,10 @@
 #include <algorithm>
 #include <cstddef>
 #include "Basics/Common.h"
-#include "Pregel/Graph/Graph.h"
+#include "Pregel/GraphStore/Graph.h"
 #include "Pregel/Worker/GraphStore.h"
-#include "Pregel/Graph/Graph.h"
-#include "Pregel/OutgoingCache.h"
 #include "Pregel/Worker/WorkerConfig.h"
+#include "Pregel/OutgoingCache.h"
 #include "Pregel/WorkerContext.h"
 
 namespace arangodb {
